Move bracket components to brackets/ folder. Add todos.
