+++ frontend/src/lib/model/Bracket.ts
@@ -1,5 +1,6 @@
 import type { Identity } from "$lib/model/Identity";
 
+// TODO: Merge / unify types in here with the types in src/lib/model/.
 export interface BracketData {
   stages: BracketStage[];
 }
