+++ frontend/src/routes/tournaments/[tournamentId]/bracket/BracketMatchNode.svelte
